backend/requirements.txt
fastapi
uvicorn
pymongo[srv]
motor
google-generativeai
python-dotenv
pydantic
python-multipart
python-dateutil
dnspython